systems: rework how systems register with theater
This changes how systems register with the theater to a component object
model. Making common how one accesses an instance of a system in the
theater and registering a system with the theater. This also forces
a common API that each system must provide:

  * must be a class and the init accepts a single argument, an instance
    of the theater
  * if the system needs to save state it just needs to implement the
    Marshallable interface

Finally this change removes the pollution of the Theater with individual
member objects, they are now all contained under `Theater.systems`.