libs: implement a generic timer class
This count up timer takes the timeout and will count the amount of time
elapsed since the last call to update(). Once elapsed time is greater
than timeout the timer has expired. The timer notes the current time
using timefunc and then takes the difference between the last time
and the current time to figure out the elapsed time since the last
call to update().

This timer does not handle roll overs of the underlying timefunc.