AssetManager: prevent removal of Player assets
Player assets are special and the code assumes they are never deleted.
They also utilize setting the dead flag to emit a DCT asset dead event
so that ticket accounting is done properly. This unfortunately makes
Player assets look dead to the rest of the system and so we need to
teach the Player assets to appear to the rest of the system that they
are not dead while keeping the nice inherited functionality of sending
a dead event.

Fixes: 07f24fd41857 ("asset manager: check if assets are dead")