SAW CAP MVP
Introduces a new auto-generated briefing style for CAP missions, and tries to keep a minimum of 4 players assigned to CAP in a region. Completing all strategic objectives in a region will cause the CAP mission in that region to also complete, rewarding players for helping others. Airspace assets now have owners based on the number of strategic assets in their regions, and once they're taken out, the airspace becomes neutral.