Add support for templates with mixed coalitions
When a template contains units of multiple coalitions, the asset's coalition must be manually specified in the .dct file, otherwise an error is thrown. This is also required in case the template doesn't have any affiliated units, such as scenery objects.

If the template doesn't have any death goals, they will only be auto-generated for units (or structures) that belong to the same coalition as the asset. This may break scenery object templates that are missing death goals, as structures no longer belong to any coalition.

This also fixes a bug where coalitions in templates could only be defined using a number, but not a string. `coalition = "blue"` and `coalition = "red"` now work as documented.

The new behavior is tested with a new template, `mixedAbrams`, which contains both a BLUFOR and a REDFOR group, but is defined as a REDFOR asset (as the objective is to eliminate only those units).